refactor and use consistent naming on dtos
