+++ apps/api/src/controllers/accounts.controller.ts
@@ -1,7 +1,7 @@
 import { Body, Controller, Get, Post, HttpCode, HttpStatus, Logger, Param, HttpException } from '@nestjs/common';
 import { ApiBody, ApiOkResponse, ApiCreatedResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { AccountResponse } from '../../../../libs/common/src/types/dtos/accounts.response.dto';
 import { AccountsService } from '../services/accounts.service';
-import { AccountResponse } from '../../../../libs/common/src/types/dtos/accounts.dto';
 import { WalletLoginResponse } from '../../../../libs/common/src/types/dtos/wallet.login.response.dto';
 import { WalletLoginRequest } from '../../../../libs/common/src/types/dtos/wallet.login.request.dto';
