+++ apps/api/src/controllers/keys.controller.ts
@@ -1,9 +1,11 @@
 import { Controller, Get, HttpCode, HttpStatus, Logger, Param, HttpException, Post, Body } from '@nestjs/common';
 import { ApiBody, ApiOkResponse, ApiOperation, ApiTags } from '@nestjs/swagger';
+import { KeysRequest, PublishKeysRequest } from '../../../../libs/common/src/types/dtos/keys.request.dto';
+import { KeysResponse } from '../../../../libs/common/src/types/dtos/keys.response.dto';
 import { KeysService } from '../services/keys.service';
-import { AddKeysRequest, KeysResponse, PublishKeysRequest } from '../../../../libs/common/src/types/dtos/keys.dto';
 import { EnqueueService } from '../../../../libs/common/src/services/enqueue-request.service';
-import { TransactionResponse, TransactionType } from '../../../../libs/common/src';
+import { TransactionType } from '../../../../libs/common/src';
+import { TransactionResponse } from '../../../../libs/common/src/types/dtos/transaction.response.dto';
 
 @Controller('keys')
 @ApiTags('keys')
@@ -21,14 +23,14 @@ export class KeysController {
   @HttpCode(HttpStatus.OK)
   @ApiOperation({ summary: 'add new control keys for an MSA ID' })
   @ApiOkResponse({ description: 'Found public keys.' })
-  @ApiBody({ type: AddKeysRequest })
+  @ApiBody({ type: KeysRequest })
   /**
    * Gets public keys.
    * @param queryParams - The query parameters for getting the public keys.
    * @returns A promise that resolves to an array of public keys associated with the given msaId.
    * @throws An error if no public keys can be found.
    */
-  async addKey(@Body() addKeysRequest: AddKeysRequest): Promise<TransactionResponse> {
+  async addKey(@Body() addKeysRequest: KeysRequest): Promise<TransactionResponse> {
     try {
       const response = await this.enqueueService.enqueueRequest<PublishKeysRequest>({
         ...addKeysRequest,
