+++ apps/api/src/services/handles.service.ts
@@ -1,13 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
-import { InjectQueue } from '@nestjs/bullmq';
-import { Queue } from 'bullmq';
 import type { HandleResponse } from '@frequency-chain/api-augment/interfaces';
-import { createHash } from 'crypto';
-import { QueueConstants } from '../../../../libs/common/src';
 import { BlockchainService } from '../../../../libs/common/src/blockchain/blockchain.service';
-import { PublishHandleRequest } from '../../../../libs/common/src/types/dtos/handles.dto';
-import { ConfigService } from '../../../../libs/common/src/config/config.service';
-import { TransactionData, TransactionResponse } from '../../../../libs/common/src/types/dtos/transaction.dto';
 
 @Injectable()
 export class HandlesService {
