+++ apps/api/src/services/keys.service.ts
@@ -1,6 +1,6 @@
 import { Injectable, Logger } from '@nestjs/common';
 import { BlockchainService } from '../../../../libs/common/src/blockchain/blockchain.service';
-import { KeysResponse } from '../../../../libs/common/src/types/dtos/keys.dto';
+import { KeysResponse } from '../../../../libs/common/src/types/dtos/keys.response.dto';
 
 @Injectable()
 export class KeysService {
