+++ libs/common/src/index.ts
@@ -4,7 +4,7 @@ export * from './utils/processing';
 export * from './utils/redis';
 export * from './utils/base-consumer';
 export * from './types/enums';
-export * from './types/dtos/transaction.dto';
+export * from './types/dtos/transaction.request.dto';
 export * from './services/nonce.service';
 export * from './utils/queues';
 export * from './services/provider-webhook.service';
