--- libs/common/src/types/dtos/account.change.notification.dto.ts
@@ -1,20 +1,0 @@
-/* eslint-disable max-classes-per-file */
-import { TransactionType } from '../enums';
-
-type Update = {
-  type: TransactionType;
-  providerId: number;
-  payload: Uint8Array;
-};
-
-export class AccountChangeNotificationDto {
-  msaId: string;
-
-  update: Update;
-}
-
-export class TransactionJob {
-  referenceId: string;
-
-  update: Update;
-}
