+++ libs/common/src/types/dtos/keys.response.dto.ts
@@ -1,0 +1,3 @@
+import { KeyInfoResponse } from '@frequency-chain/api-augment/interfaces';
+
+export type KeysResponse = KeyInfoResponse['msa_keys'];
