+++ libs/common/src/types/dtos/transaction.response.dto.ts
@@ -1,0 +1,3 @@
+export type TransactionResponse = {
+  referenceId: string;
+};
