+++ libs/common/src/types/dtos/wallet.login.response.dto.ts
@@ -1,5 +1,4 @@
 /* eslint-disable max-classes-per-file */
-import { EncodedExtrinsic } from '@amplica-labs/siwf';
 import { IsNotEmpty, IsOptional } from 'class-validator';
 import { TransactionType } from '../enums';
